@@ -1,7 +1,7 @@
 package br.ronanlima.opiniaodetudo
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import br.ronanlima.opiniaodetudo.view.ListFragment
 
 class ListaActivity : AppCompatActivity() {
