@@ -1,11 +1,11 @@
 package br.ronanlima.opiniaodetudo.infra.dao.room
 
-import android.arch.persistence.db.SupportSQLiteDatabase
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
-import android.arch.persistence.room.migration.Migration
 import android.content.Context
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
+import androidx.room.migration.Migration
+import androidx.sqlite.db.SupportSQLiteDatabase
 import br.ronanlima.opiniaodetudo.infra.dao.ReviewTableInfo
 import br.ronanlima.opiniaodetudo.model.Review
 
